Refactor tests structure to split tests library
