+++ roles/oneagent/tests/command/platform_command_wrapper.py
@@ -1,9 +1,10 @@
 from pathlib import Path
 
+from command.command_result import CommandResult
 from command.command_wrapper import CommandWrapper
 from command.unix.unix_command_wrapper import UnixCommandWrapper
 from command.windows.windows_command_wrapper import WindowsCommandWrapper
-from util.test_data_types import CommandResult, DeploymentPlatform
+from deployment.deployment_platform import DeploymentPlatform
 
 
 class PlatformCommandWrapper:
