+++ roles/oneagent/tests/command/windows/windows_command_executor.py
@@ -1,5 +1,5 @@
 import winrm
-from util.test_data_types import CommandResult
+from command.command_wrapper import CommandResult
 
 
 class WindowsCommandExecutor:
