+++ roles/oneagent/tests/deployment/deployment_operations.py
@@ -1,22 +1,94 @@
 import functools
 import logging
+import os
+import shutil
 from pathlib import Path
-from typing import Any, Callable
+from typing import Any, Callable, Dict
 
 from ansible.config import AnsibleConfig
 from ansible.runner import AnsibleRunner
 from command.platform_command_wrapper import PlatformCommandWrapper
-from util.common_utils import get_oneagentctl_path, get_platform_argument
 from constants import (
     INSTALLER_CERTIFICATE_FILE_NAME,
     INSTALLER_PARTIAL_NAME,
     INSTALLER_SERVER_TOKEN,
+    INSTALLER_SYSTEM_NAME_TYPE_MAP,
     SERVER_CERTIFICATE_FILE_NAME,
+    UNIX_ONEAGENTCTL_PATH,
+    WINDOWS_ONEAGENTCTL_PATH,
+    WORK_DIR_PATH,
+    WORK_INSTALLERS_DIR_PATH,
     WORK_SERVER_DIR_PATH,
 )
-from util.test_data_types import DeploymentPlatform, DeploymentResult, PlatformCollection
+
+from .deployment_platform import (
+    DeploymentPlatform,
+    DeploymentResult,
+    PlatformCollection,
+)
 
 CallableOperation = Callable[[DeploymentPlatform, str, Any], None]
+
+
+def prepare_test_dirs() -> None:
+    remove_if_exists(WORK_DIR_PATH)
+    Path(WORK_DIR_PATH).mkdir(parents=True)
+    os.chdir(WORK_DIR_PATH)
+
+
+def remove_if_exists(path: Path) -> None:
+    if path.exists():
+        try:
+            shutil.rmtree(str(path)) if os.path.isdir(str(path)) else os.remove(str(path))
+        except OSError as os_error:
+            logging.error("Failed to remove %s: %s", path, os_error)
+
+
+def get_oneagentctl_path(platform: DeploymentPlatform) -> Path:
+    return get_platform_argument(platform, UNIX_ONEAGENTCTL_PATH, WINDOWS_ONEAGENTCTL_PATH)
+
+
+def get_platform_argument(platform: DeploymentPlatform, unix_arg: Any, windows_arg: Any) -> Any:
+    return windows_arg if platform == DeploymentPlatform.WINDOWS_X86 else unix_arg
+
+
+def _get_platform_by_installer(installer: Path) -> DeploymentPlatform:
+    name = installer.name.lower()
+    for platform in DeploymentPlatform:
+        if platform.arch() in name and platform.system() in name:
+            return platform
+
+    # Special handling for Linux_x86 and Windows as the installer does not
+    # contain architecture in its name
+    if DeploymentPlatform.WINDOWS_X86.system() in name:
+        return DeploymentPlatform.WINDOWS_X86
+    return DeploymentPlatform.LINUX_X86
+
+
+def _get_available_installers() -> Dict[DeploymentPlatform, list[Path]]:
+    installers: dict[DeploymentPlatform, list[Path]] = {k: [] for k in DeploymentPlatform}
+    for installer in sorted(WORK_INSTALLERS_DIR_PATH.glob(f"{INSTALLER_PARTIAL_NAME}*")):
+        platform = _get_platform_by_installer(installer)
+        installers[platform].append(installer)
+    return installers
+
+
+def get_installers(system: str, arch: str, version: str = "", include_paths: bool = False) -> list[Path | str]:
+    try:
+        # Special handling for mocking server behavior as URL for Linux
+        # installers contains "unix" instead of linux
+        system = INSTALLER_SYSTEM_NAME_TYPE_MAP[system]
+        platform_installers = _get_available_installers()[DeploymentPlatform.from_system_and_arch(system, arch)]
+        installers = platform_installers if include_paths else [ins.name for ins in platform_installers]
+        if not version:
+            return installers
+        if version == "latest":
+            return [installers[-1]]
+        return [installer for installer in installers if version in str(installer)]
+
+    except Exception as ex:
+        logging.error("Failed to get installer for %s_%s in %s version: %s", system, arch, version, ex)
+        return []
 
 
 def _get_param_by_name(name: str, **kwargs) -> Any:
@@ -51,7 +123,9 @@ def set_ca_cert_download_params(config: AnsibleConfig, installer_server_url: str
     config.set_common_parameter(
         config.CA_CERT_DOWNLOAD_URL_KEY, f"{installer_server_url}/{INSTALLER_CERTIFICATE_FILE_NAME}"
     )
-    config.set_common_parameter(config.CA_CERT_DOWNLOAD_CERT_KEY, f"{WORK_SERVER_DIR_PATH / SERVER_CERTIFICATE_FILE_NAME}")
+    config.set_common_parameter(
+        config.CA_CERT_DOWNLOAD_CERT_KEY, f"{WORK_SERVER_DIR_PATH / SERVER_CERTIFICATE_FILE_NAME}"
+    )
     config.set_common_parameter(config.FORCE_CERT_DOWNLOAD_KEY, True)
