+++ roles/oneagent/tests/test_install_and_config.py
@@ -2,16 +2,21 @@ import logging
 from pathlib import Path
 
 from command.platform_command_wrapper import PlatformCommandWrapper
-from util.common_utils import get_oneagentctl_path, get_platform_argument
-from constants import INSTALLER_SERVER_TOKEN, UNIX_DOWNLOAD_DIR_PATH, WINDOWS_DOWNLOAD_DIR_PATH
-from util.test_data_types import DeploymentPlatform, DeploymentResult
-from util.test_helpers import (
+from constants import (
+    INSTALLER_SERVER_TOKEN,
+    UNIX_DOWNLOAD_DIR_PATH,
+    WINDOWS_DOWNLOAD_DIR_PATH,
+)
+from deployment.deployment_operations import (
     check_agent_state,
     check_download_directory,
+    get_oneagentctl_path,
+    get_platform_argument,
     perform_operation_on_platforms,
     run_deployment,
     set_installer_download_params,
 )
+from deployment.deployment_platform import DeploymentPlatform, DeploymentResult
 
 CTL_OPTION_GET_HOST_TAGS = "--get-host-tags"
 CTL_OPTION_SET_HOST_TAG = "--set-host-tag"
