--- roles/oneagent/tests/util/constant_key.py
@@ -1,5 +1,0 @@
-from abc import abstractmethod
-
-
-def constant_key(func) -> None:
-    property(classmethod(abstractmethod(func)))
